增加获取系统时间的方法

+++ mamba-framework-context/src/main/java/com/mamba/framework/context/session/core/Operator.java
@@ -3,5 +3,5 @@ package com.mamba.framework.context.session.core;
 import java.io.Serializable;
 
 public interface Operator extends Serializable {
-	public long getOperatorId();
+	public int getOperatorId();
 }
