+++ mamba-framework-context/src/main/java/com/mamba/framework/context/session/provider/OperatorProvider.java
@@ -3,5 +3,5 @@ package com.mamba.framework.context.session.provider;
 import com.mamba.framework.context.session.core.Operator;
 
 public interface OperatorProvider {
-	public Operator getOperator(long operatorId);
+	public Operator getOperator(int operatorId);
 }
